AppliedPredictiveModeling-package {AppliedPredictiveModeling}	R Documentation

Data, Functions and Scripts for 'scriptLocation'

Description

This package can be used to reproduce the analyses in the text. Scripts for each chapter are located in the "chapters" directory. Use scriptLocation() to find their exact location.

Details

Package: 	 AppliedPredictiveModeling

Type: 	 Package

Version: 	 1.1-1

Date: 	 2013-05-29

License: 	 GPL

Author(s)

Max Kuhn

Maintainer: Max Kuhn <mkuhn@gmail.com>

References

Kuhn M and Johnson K (2013) Applied Predictive Modeling, Springer, NY

[Package AppliedPredictiveModeling version 1.1-6 Index]