bookTheme {AppliedPredictiveModeling}	R Documentation

Lattice Themes

Description

Two lattice themes used throughout the book.

Usage

bookTheme(set = TRUE)

transparentTheme(set = TRUE, pchSize = 1, trans = 0.2)

Arguments

set
a logical: should these settings be applied to the current device?

pchSize
the size of the plot symbols

trans
the amount of transparency (via the alpha channel). Note that transparency is not supported by all graphics devices.

Details

When using these functions to save a plot, make sure to invoke them after the device has been opened (e.g. after calls such as pdf().

Value

Each function returns a list of theme parameters. See Sarkar (2008) or trellis.par.get for specific details.

Author(s)

Max Kuhn

References

Some of the colors are based on values from ColorBrewer http://www.colorbrewer.org.

Sarkar, D. (2008). Lattice: Multivariate Data Visualization with R. UseR! (1st ed. p. 286). Springer.

Examples

library(lattice)

example <- quadBoundaryFunc(100)

bookTheme(set = TRUE)
xyplot(X2 ~ X1, data = example, groups = class, auto.key = TRUE)

transparentTheme(set = TRUE, trans = .6)
xyplot(X2 ~ X1, data = example, groups = class, auto.key = TRUE)

[Package AppliedPredictiveModeling version 1.1-6 Index]