getPackages {AppliedPredictiveModeling}	R Documentation

Install Packages for Each Chapter

Description

This function identifies the physical location on the user's computer where the chapter R scripts are located.

Usage

getPackages(chapter, ...)

Arguments

chapter
an integer vector (or character versions of the integer) for the chapter number. See Details below:

...
options to pass to install.packages

Details

Chapter names and packages. about dependencies.

Author(s)

Max Kuhn

Examples

## Not run: 
getPackages(2)
getPackages(2:3)
getPackages("4")

## End(Not run)

[Package AppliedPredictiveModeling version 1.1-6 Index]