scriptLocation {AppliedPredictiveModeling}	R Documentation

Find Chapter Script Files

Description

This function identifies the physical location on the user's computer where the chapter R scripts are located.

Usage

scriptLocation()

Author(s)

Max Kuhn

Examples

scriptLocation()

[Package AppliedPredictiveModeling version 1.1-6 Index]